Fixing server hot reloading
