+++ data/config/server-config-default.yaml
@@ -10,3 +10,4 @@ showWelcome: true
 expRate: 1
 giveAchievements: true
 checkCredentials: true
+clippingDisabled: false
