+++ src/net/server/server-config.ts
@@ -1,0 +1,15 @@
+export interface ServerConfig {
+    rsaMod: string;
+    rsaExp: string;
+    host: string;
+    port: number;
+    loginServerHost: string;
+    loginServerPort: number;
+    updateServerHost: string;
+    updateServerPort: number;
+    showWelcome: boolean;
+    expRate: number;
+    giveAchievements: boolean;
+    checkCredentials: boolean;
+    clippingDisabled: boolean;
+}
