+++ src/util/files.ts
@@ -1,6 +1,7 @@
 import util from 'util';
 import fs from 'fs';
 import { watch } from 'chokidar';
+import { Observable, Subject } from 'rxjs';
 
 const readdir = util.promisify(fs.readdir);
 const stat = util.promisify(fs.stat);
@@ -26,6 +27,18 @@ export async function* getFiles(directory: string, blacklist: string[]): AsyncGe
             yield path;
         }
     }
+}
+
+export function watchSource(dir: string): Observable<void> {
+    const subject = new Subject<void>();
+    const watcher = watch(dir);
+    watcher.on('ready', () => {
+        watcher.on('all', () => {
+            subject.next();
+        });
+    });
+
+    return subject.asObservable();
 }
 
 export function watchForChanges(dir: string, regex: RegExp): void {
